add recycler

+++ app/src/main/java/by/ocheretny/homework/homework9/data/dto/crypto/CryptoResponse.kt
@@ -11,7 +11,7 @@ data class CryptoResponse(
 ) {
     data class Data(
         @SerializedName("circulating_supply")
-        val circulatingSupply: Int?,
+        val circulatingSupply: Double?,
         @SerializedName("cmc_rank")
         val cmcRank: Int?,
         @SerializedName("date_added")
@@ -21,7 +21,7 @@ data class CryptoResponse(
         @SerializedName("last_updated")
         val lastUpdated: String?,
         @SerializedName("max_supply")
-        val maxSupply: Int?,
+        val maxSupply: Double?,
         @SerializedName("name")
         val name: String?,
         @SerializedName("num_market_pairs")
@@ -37,7 +37,7 @@ data class CryptoResponse(
         @SerializedName("tags")
         val tags: List<String?>?,
         @SerializedName("total_supply")
-        val totalSupply: Int?
+        val totalSupply: Double?
     ) {
         data class Quote(
             @SerializedName("USD")
